@@ -11,13 +11,13 @@
 #include <vector>
 #include "CVContourLineExtractionAlgorithm.hpp"
 
-void mmk::CVContourLineExtractionAlgorithm::extract(const mmk::Image& input, mmk::Image*const output)const
+void mmk::CVContourLineExtractionAlgorithm::execute(const mmk::Image& input, mmk::Image*const output)
 {
     std::vector<std::vector<cv::Point> > contours;
     const mmk::OpenCVImage& cvImage = static_cast<const mmk::OpenCVImage&>(input);
     mmk::OpenCVImage*const cvOutputImage = static_cast<mmk::OpenCVImage*const>(output);
     cv::Mat grayImage;
-    cvtColor(cvImage, grayImage, CV_BGR2GRAY);
+    cvtColor(cvImage, grayImage, cv::COLOR_RGB2GRAY);
 
     cv::findContours(grayImage, contours, cv::RetrievalModes::RETR_LIST, cv::ContourApproximationModes::CHAIN_APPROX_NONE);
     
