@@ -9,19 +9,19 @@
 #ifndef ContourLineExtractionAlgorithm_h
 #define ContourLineExtractionAlgorithm_h
 
+#include "effect/EffectComponent.hpp"
+
 namespace mmk {
     class Image;
     class ContourLineExtractionAlgorithm;
 }
 
-class mmk::ContourLineExtractionAlgorithm
+class mmk::ContourLineExtractionAlgorithm : public mmk::EffectComponent
 {
 protected:
     ContourLineExtractionAlgorithm() = default;
 public:
     virtual~ContourLineExtractionAlgorithm() = default;
-    
-    virtual void extract(const mmk::Image& input, mmk::Image*const output)const = 0;
 };
 
 #endif /* ContourLineExtractionAlgorithm_h */
