@@ -10,14 +10,18 @@
 #define EffectComponent_hpp
 
 namespace mmk {
+    class Image;
     class EffectComponent;
 }
 
 class mmk::EffectComponent
 {
-public:
+protected:
     EffectComponent() = default;
+public:
     virtual~EffectComponent() = default;
+    
+    virtual void execute(const mmk::Image& input, mmk::Image*const output) = 0;
 };
 
 #endif /* EffectComponent_hpp */
