@@ -6,13 +6,13 @@
 //
 //
 
-
+#include <iostream>
 #include "image/OpenCVImage.hpp"
 #include "OpenCVVideoCapture.hpp"
 
 bool mmk::OpenCVVideoCapture::initialize()
 {
-    m_videoCapture->set(CV_CAP_PROP_FPS, 24);
+    m_videoCapture->set(CV_CAP_PROP_FPS, 20);
     m_videoCapture->set(CV_CAP_PROP_FRAME_WIDTH, 640);
     m_videoCapture->set(CV_CAP_PROP_FRAME_HEIGHT, 480);
     
@@ -22,7 +22,7 @@ bool mmk::OpenCVVideoCapture::initialize()
 void mmk::OpenCVVideoCapture::capture(mmk::Image *image)
 {
     mmk::OpenCVImage* cvImage = static_cast<mmk::OpenCVImage*>(image);
-    (*getVideoCapturePtr()) >> (*cvImage);
+    *getVideoCapturePtr() >> *cvImage;
 }
 
 cv::VideoCapture*const mmk::OpenCVVideoCapture::getVideoCapturePtr()
